pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Andrew Klyuchnikov <off-rey@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.1"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.5"
django-cors-headers = "^4.0.0"
pytest = "^7.3.1"
pytest-django = "^4.5.2"
pyyaml = "^6.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]
pre-commit = "^3.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
